pyproject.toml
[tool.poetry]
name = "ami-ml"
version = "0.1.0"
description = "Software & research related to the automated monitoring of insects"
authors = [
    "Aditya Jain <aditya.jain@mila.quebec>",
    "Michael Bunsen <michael.bunsen@mila.quebec>",
    "Fagner Cunha <fagner.cunha@icomp.ufam.edu.br>",
    "Léonard Pasi <leonardpasi@gmail.com>",
]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.6"
numpy = "^1.25.1"
torch = "^2.0,!=2.0.1"
tqdm = "^4.65.0"
segment-anything = "^1.0"
torchvision = "^0.15.1,!=0.15.2"
torchmetrics = "^0.11.4"
pillow = "^10.0.0"
matplotlib = "^3.7.2"
wandb = "^0.14.0"
python-dwca-reader = "^0.15.1"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
webdataset = "^0.2.48"
timm = "^0.9.8"
typer = "^0.12.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = "^23.7.0"
flake8 = "^6.0.0"
Flake8-pyproject = "^1.2.3"
isort = "^5.12.0"
mypy = "^1.4.1"
ipdb = "^0.13.13"
python-devtools = "^2"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
